move BusDto to snu-lib
