+++ admin/src/scenes/plan-transport/ligne-bus/View/components/Info.tsx
@@ -4,12 +4,12 @@ import { toastr } from "react-redux-toastr";
 import { Link } from "react-router-dom";
 
 import { canEditLigneBusGeneralInfo, isBusEditionOpen, translate } from "snu-lib";
+import { BusDto } from "snu-lib/src/dto";
 import { Button } from "@snu/ds/admin";
 
 import { capture } from "@/sentry";
 import { AuthState } from "@/redux/auth/reducer";
 import api from "@/services/api";
-import { Bus } from "@/types";
 import Pencil from "@/assets/icons/Pencil";
 import Loader from "@/components/Loader";
 
@@ -24,7 +24,7 @@ const options = [
 ];
 
 interface Props {
-  bus: Bus;
+  bus: BusDto;
   onBusChange: (bus: Props["bus"]) => void;
   dataForCheck: {
     youngsCountBus: number;
