+++ admin/src/types.ts
@@ -48,34 +48,3 @@ export type Session = {
   canBeDeleted?: boolean;
   sanitaryContactEmail: string;
 };
-
-export type Bus = {
-  _id?: string;
-  cohort?: string;
-  busId?: string;
-  departuredDate?: string;
-  returnDate?: string;
-  youngCapacity?: number | string;
-  youngSeatsTaken?: number | string;
-  totalCapacity?: number | string;
-  followerCapacity?: number | string;
-  travelTime?: string;
-  lunchBreak?: string;
-  lunchBreakReturn?: string;
-  classeId?: string;
-  mergedBusIds?: string[];
-  mergedBusDetails?: Pick<Bus, "_id" | "busId" | "totalCapacity" | "youngCapacity" | "youngSeatsTaken">[];
-  team?: BusTeam[];
-};
-
-export interface BusTeam {
-  _id?: string;
-  role?: string;
-  lastName?: string;
-  firstName?: string;
-  birthdate?: Date;
-  mail?: string;
-  phone?: string;
-  forth?: boolean;
-  back?: boolean;
-}
