+++ packages/lib/src/dto/index.ts
@@ -1,2 +1,3 @@
 export * from "./cohortDto";
 export * from "./classeDto";
+export * from "./busDto";
